requirements.txt

# python==3.8.19
numpy==1.24.4
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.3.2
matplotlib==3.7.5


einops==0.8.0
pytorch-lightning==2.4.0
tensorboard==2.14.0
torchmetrics==1.5.2

opencv-python==4.6.0.66
imageio==2.35.1
scikit-image==0.21.0
pywavelets==1.4.1

transformers==4.46.3
huggingface-hub==0.28.1


timm==1.0.14
omegaconf==2.3.0
segment-anything==1.0
pretrainedmodels==0.7.4
grad-cam==1.4.3

tqdm==4.66.6
requests==2.32.3
protobuf==5.28.3
six==1.16.0

# conda install pytorch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 pytorch-cuda=11.8 -c pytorch -c nvidia